requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
azure-ai-ml==1.12.1
azure-common==1.1.28
azure-core==1.29.6
azure-graphrbac==0.61.1
azure-identity==1.15.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-network==25.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.15.0
azureml-automl-core==1.54.0
azureml-core==1.54.0.post1
azureml-dataprep==4.12.8
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.19.6
azureml-dataset-runtime==1.54.0.post1
azureml-pipeline==1.54.0
azureml-pipeline-core==1.54.0
azureml-pipeline-steps==1.54.0
azureml-sdk==1.54.0
azureml-telemetry==1.54.0
azureml-train-automl-client==1.54.0.post1
azureml-train-core==1.54.0
azureml-train-restclients-hyperdrive==1.54.0
gevent==23.9.1
geventhttpclient==2.0.2
httptools==0.6.1
multiprocess==0.70.15
pyOpenSSL==23.3.0
python-dotenv==1.0.0
safetensors==0.4.1
sentencepiece==0.1.99
tokenizers==0.15.0
torch==2.1.2
torchsummary==1.5.1
torchvision==0.16.2
transformers==4.36.0
triton==2.1.0
tritonclient==2.41.1
xformers==0.0.23
